+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2025.04.06
 - fix issue with not changing unitTags
 - reformatting code structure
+- fire player events on group update
 
 ## 2025.03.12
 - dump dependencies: